pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["bashar adel taamneh <bashartaamneh55@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
black = "^21.11b1"
flake8 = "^4.0.1"
asgiref = "^3.4.1"
django-cors-headers = "^3.10.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
PyJWT = "^2.3.0"
pytz = "^2021.3"
sqlparse = "^0.4.2"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.11b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
